Move codec-related classes into dedicated subpackages (base64, frame, json)
