+++ core/src/main/java/org/traffichunter/titan/core/codec/base64/Base64Codec.java
@@ -21,37 +21,48 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.traffichunter.titan.core.codec;
+package org.traffichunter.titan.core.codec.base64;
 
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
+import java.nio.ByteBuffer;
+import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
 
 /**
  * @author yungwang-o
  */
-public abstract class Headers<K, V, H extends Headers<K, V, H>> {
+public final class Base64Codec {
 
-    protected final Map<K, V> map;
+    private static final Base64.Encoder BASE64_ENCODER = Base64.getEncoder().withoutPadding();
+    private static final Base64.Decoder BASE64_DECODER = Base64.getDecoder();
 
-    protected Headers(final Map<K, V> map) {
-        this.map = map;
+    public static byte[] encode(final byte[] src) {
+        return BASE64_ENCODER.encode(src);
     }
 
-    public abstract void put(K key, V value);
+    public static byte[] encode(final String src) {
+        return encode(src, StandardCharsets.UTF_8);
+    }
 
-    public abstract void putIfAbsent(K key, V value);
+    public static byte[] encode(final String src, final Charset charset) {
+        return BASE64_ENCODER.encode(src.getBytes(charset));
+    }
 
-    public abstract Optional<V> get(K key);
+    public static byte[] encode(final ByteBuffer src) {
+        return BASE64_ENCODER.encode(src).array();
+    }
 
-    public abstract boolean containsKey(K key);
+    public static byte[] decode(final byte[] src) {
+        return BASE64_DECODER.decode(src);
+    }
 
-    public abstract Set<K> keySet();
+    public static byte[] decode(final String src) {
+        return BASE64_DECODER.decode(src);
+    }
 
-    public abstract Set<Map.Entry<K, V>> entrySet();
+    public static byte[] decode(final ByteBuffer src) {
+        return BASE64_DECODER.decode(src).array();
+    }
 
-    public abstract Iterator<Map.Entry<K, V>> iterator();
-
-    public abstract H getHeader();
+    private Base64Codec() { }
 }
